+++ fingerprint/infra/nec-bio-sdk/src/main/java/com/simprints/fingerprint/infra/necsdkimpl/NecSdkModule.kt
@@ -28,7 +28,7 @@ object NecSdkModule {
 
     @Provides
     @Singleton
-    fun provideFingerprintBioSdk(
+    internal fun provideFingerprintBioSdk(
         sdkInitializer: SdkInitializerImpl,
         fingerprintImageProvider: FingerprintImageProviderImpl,
         fingerprintTemplateProvider: FingerprintTemplateProviderImpl,
