+++ fingerprint/infra/nec-bio-sdk/src/main/java/com/simprints/fingerprint/infra/necsdkimpl/acquisition/image/FingerprintImageProviderImpl.kt
@@ -5,7 +5,7 @@ import com.simprints.fingerprint.infra.basebiosdk.acquisition.domain.ImageRespon
 import com.simprints.fingerprint.infra.basebiosdk.exceptions.BioSdkException
 import javax.inject.Inject
 
-class FingerprintImageProviderImpl @Inject constructor(
+internal class FingerprintImageProviderImpl @Inject constructor(
     private val imageCache: ProcessedImageCache
 ) : FingerprintImageProvider<Unit, Unit> {
