+++ fingerprint/infra/nec-bio-sdk/src/test/java/com/simprints/fingerprint/infra/necsdkimpl/acquisition/image/FingerprintImageProviderImplTest.kt
