+++ fingerprint/infra/simprints-bio-sdk/src/main/java/com/simprints/fingerprint/infra/biosdkimpl/acquisition/image/FingerprintImageProviderImpl.kt
@@ -5,7 +5,7 @@ import com.simprints.fingerprint.infra.basebiosdk.acquisition.domain.ImageRespon
 import com.simprints.fingerprint.infra.scanner.capture.FingerprintCaptureWrapperFactory
 import javax.inject.Inject
 
-class FingerprintImageProviderImpl @Inject constructor(
+internal class FingerprintImageProviderImpl @Inject constructor(
     private val fingerprintCaptureWrapperFactory: FingerprintCaptureWrapperFactory
 ) : FingerprintImageProvider<Unit, Unit> {
