+++ fingerprint/infra/simprints-bio-sdk/src/main/java/com/simprints/fingerprint/infra/biosdkimpl/initialization/SdkInitializerImpl.kt
@@ -3,7 +3,7 @@ package com.simprints.fingerprint.infra.biosdkimpl.initialization
 import com.simprints.fingerprint.infra.basebiosdk.initialization.SdkInitializer
 import javax.inject.Inject
 
-class SdkInitializerImpl @Inject constructor() : SdkInitializer<Unit> {
+internal class SdkInitializerImpl @Inject constructor() : SdkInitializer<Unit> {
     override suspend fun initialize(initializationParams: Unit?) {
         // No need to initialize anything for the Simprints Bio SDK
     }
