+++ fingerprint/infra/simprints-bio-sdk/src/main/java/com/simprints/fingerprint/infra/biosdkimpl/matching/FingerprintMatcherImpl.kt
@@ -1,12 +1,11 @@
 package com.simprints.fingerprint.infra.biosdkimpl.matching
 
 import com.simprints.fingerprint.infra.basebiosdk.matching.FingerprintMatcher
-import com.simprints.fingerprint.infra.basebiosdk.matching.SimAfisMatcher
 import com.simprints.fingerprint.infra.basebiosdk.matching.domain.FingerprintIdentity
 import com.simprints.fingerprint.infra.basebiosdk.matching.domain.MatchResult
 import javax.inject.Inject
 
-class FingerprintMatcherImpl @Inject constructor(
+internal class FingerprintMatcherImpl @Inject constructor(
     private val simAfisMatcher: SimAfisMatcher
 ) : FingerprintMatcher<SimAfisMatcherSettings> {
 
@@ -17,4 +16,4 @@ class FingerprintMatcherImpl @Inject constructor(
     ): List<MatchResult> = simAfisMatcher.match(probe, candidates, settings?.crossFingerComparison?:false)
 }
 
-data class SimAfisMatcherSettings(var crossFingerComparison: Boolean = false)
+
