+++ fingerprint/infra/simprints-bio-sdk/src/main/java/com/simprints/fingerprint/infra/biosdkimpl/matching/SimAfisMatcherSettings.kt
@@ -1,0 +1,3 @@
+package com.simprints.fingerprint.infra.biosdkimpl.matching
+
+data class SimAfisMatcherSettings(var crossFingerComparison: Boolean = false)
